scripts/blackjack/requirements.txt
pillow>=10.0
pytesseract>=0.3.10
opencv-python-headless>=4.8
numpy>=1.24
rich>=13.0
websockets>=12.0
